+++ .github/workflows/ci-go-cover.yml
@@ -31,6 +31,5 @@ jobs:
     - name: Go Coverage
       run: |
         go version
-        make GOTEST_ARGS="-short -cover" test | grep -o "coverage:.*of statements$" | python scripts/cov.py
-
+        make test-cover | grep -o "coverage:.*of statements$" | python scripts/cov.py
       shell: bash
