+++ tpm/kat_test.go
@@ -1,6 +1,6 @@
 // Copyright 2022 Contributors to the Veraison project.
 // SPDX-License-Identifier: Apache-2.0
-package parsectpm
+package tpm
 
 import (
 	"crypto/ecdsa"
