+++ tpm/test_common.go
@@ -1,7 +1,7 @@
 // Copyright 2023 Contributors to the Veraison project.
 // SPDX-License-Identifier: Apache-2.0
 
-package parsectpm
+package tpm
 
 import (
 	"crypto"
